+++ src/components/posts/GetPost.tsx
@@ -9,7 +9,7 @@ const GetPost = () => {
   const [posts, setPosts] = useState<Post[]>([]);
 
   useEffect(() => {
-    async function fetchPosts() {
+    const fetchPosts = async () => {
       const { data, error } = await supabase.from('posts').select('*');
 
       if (error) {
@@ -22,7 +22,7 @@ const GetPost = () => {
 
         setPosts(postsWithCompleteURLs);
       }
-    }
+    };
 
     fetchPosts();
   }, []);
